+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Add `add` and `autoadd` method.
 - Migrate `logLevel` parameter to global.
 
+**Helm**
+- Add `dependency` method.
+
 **Packer**
 - Increase minimum supported version to 1.9.
 - Add `machineReadable`, `parallelBuilds`, `timestampUi`, `warnUndeclaredVar`, `ignorePrerelease`, and `sequentialEval` params to `build` method.